+++ tests/phpunit/WC_Stripe_Helper_Test.php
@@ -896,25 +896,4 @@ class WC_Stripe_Helper_Test extends WC_Mock_Stripe_API_Unit_Test_Case {
 			'webhook URL with extra parameters should match'           => [ 'https://example.com/test/?wc-api=wc_stripe&foo=bar', 'https://example.com/test/?wc-api=wc_stripe', true ],
 		];
 	}
-
-	/**
-	 * Tests for `is_stripe_gateway_order`.
-	 *
-	 * @return void
-	 */
-	public function test_is_stripe_gateway_order() {
-		// Test with a Stripe order (Klarna).
-		$order = WC_Helper_Order::create_order();
-		$order->set_payment_method( 'stripe_klarna' );
-		$this->assertTrue( WC_Stripe_Helper::is_stripe_gateway_order( $order ) );
-
-		// Test with a non-Stripe order.
-		$order = WC_Helper_Order::create_order();
-		$order->set_payment_method( 'cod' );
-		$this->assertFalse( WC_Stripe_Helper::is_stripe_gateway_order( $order ) );
-
-		// Test with an empty order.
-		$order = new WC_Order();
-		$this->assertFalse( WC_Stripe_Helper::is_stripe_gateway_order( $order ) );
-	}
 }
