+++ tests/phpunit/WC_Stripe_Order_Helper_Test.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace WooCommerce\Stripe\Tests;
+
+use WC_Stripe_Order_Helper;
+use WooCommerce\Stripe\Tests\Helpers\WC_Helper_Order;
+
+/**
+ * Class WC_Stripe_Order_Handler tests.
+ */
+class WC_Stripe_Order_Handler_Test extends WP_UnitTestCase {
+	/**
+	 * Tests for `is_stripe_gateway_order`.
+	 *
+	 * @return void
+	 */
+	public function test_is_stripe_gateway_order() {
+		// Test with a Stripe order (Klarna).
+		$order = WC_Helper_Order::create_order();
+		$order->set_payment_method( 'stripe_klarna' );
+		$this->assertTrue( WC_Stripe_Order_Helper::is_stripe_gateway_order( $order ) );
+
+		// Test with a non-Stripe order.
+		$order = WC_Helper_Order::create_order();
+		$order->set_payment_method( 'cod' );
+		$this->assertFalse( WC_Stripe_Order_Helper::is_stripe_gateway_order( $order ) );
+
+		// Test with an empty order.
+		$order = new WC_Order();
+		$this->assertFalse( WC_Stripe_Order_Helper::is_stripe_gateway_order( $order ) );
+	}
+}
